fix(core): text-plan properties, PG18 JSON rows, and numeric correctness
**Text plans lost their multi-field properties.** PostgreSQL packs several
fields onto one line — `Sort Method: external merge  Disk: 524288kB`,
`Heap Blocks: exact=100 lossy=900`, `Buckets: 1024  Batches: 4  Memory Usage:
44kB` — and the generic first-colon split put the whole tail into the first
key. So `Sort Space Used`, `Sort Space Type`, `Batches` and
`Heap Blocks: lossy` never existed on text plans, which is the format
auto_explain emits by default.

The user-visible consequence: SortMemoryAnalyzer reported every disk spill as
"using 0 kB", and because its severity escalation keys off the spill size, a
512 MB spill was filed as Medium. Verified end-to-end — the same plan now
reports "using 524288 kB" at High. The hash-batch and lossy-bitmap rules were
outright dead.

The property vocabulary was also written against text semantics, so
PostgreSQL's JSON spellings (`Exact Heap Blocks`, `Lossy Heap Blocks`,
`Hash Batches`) matched nothing either — those rules were dead in *both*
formats. Both spellings now canonicalise to the same keys.

**PG18 JSON plans vanished.** PG18 prints `Actual Rows` as a per-loop average
with decimals once `loops > 1`. Deserializing into `u64` made serde_json reject
the whole document, which the state machine demoted to plain query text — so on
PG18 with `log_format = json`, every plan containing a nested-loop node
disappeared from the analysis with only a warning. The text path already
accepted fractions; the two now agree.

**Numeric correctness:**

* `calculate_group_duration_stats` used the population variance divisor (N)
  while documenting itself as bit-identical to `calculate_mean_and_std_dev`,
  which uses the sample divisor (N-1). The group value is the one that reaches
  `QueryGroupStatistics.std_dev_ms` and the export, and small groups — what
  slow queries typically form — were understated by up to ~30%.
* `basic_regression` divided by the first-half average with no zero guard,
  reporting `+inf%` as a Critical regression at maximum confidence (the
  trend-strength clamp returns 1.0 for NaN because `f64::min` propagates the
  non-NaN operand). `auto_explain.log_min_duration = 0` makes all-zero halves
  routine. The statistical path already guarded this.
* The heap-fetch ratio compared cumulative `Heap Fetches` against per-loop
  `rows`, overstating it by the loop count on the inner side of every nested
  loop.

**Exports are now reproducible.** Equal-duration groups came out in the
iteration order of a randomly-seeded `hashbrown::HashMap`, so two runs over the
same input produced byte-different JSON. Ties break on query hash, and the
comparison is now total (`total_cmp`) rather than `partial_cmp().unwrap_or`.

+++ crates/core/src/models.rs
@@ -112,8 +112,15 @@ pub struct JsonPlanNode {
     pub actual_startup_time: Option<f64>,
     #[serde(rename = "Actual Total Time")]
     pub actual_total_time: Option<f64>,
+    /// PostgreSQL 18 emits this as a fraction when `loops > 1` — it prints the
+    /// per-loop *average* (e.g. `1000.5`). Deserializing into an integer made
+    /// `serde_json` reject the whole document, which the state machine then
+    /// demoted to plain query text: on a PG18 server with
+    /// `auto_explain.log_format = json`, every plan containing a node with
+    /// `loops > 1` disappeared from the analysis entirely, with only a warning.
+    /// The text parser already accepted fractions; this keeps the two in step.
     #[serde(rename = "Actual Rows")]
-    pub actual_rows: Option<u64>,
+    pub actual_rows: Option<f64>,
     #[serde(rename = "Actual Loops")]
     pub actual_loops: Option<u32>,
